+++ frontend/src/routes/food-scan/+page.svelte
@@ -7,11 +7,11 @@
   import { setLocale } from '$lib/util/translations';
   import { onMount } from 'svelte';
 
-  // 2. CONST (non-reactive constants)
-  const { 'food-scan': foodScanRoute } = applicationRoutes;
-
   // 4. STATE
   let locale = $state($languageStore);
+
+  // 5. DERIVED
+  const foodScanRoute = $derived($applicationRoutes['food-scan']);
 
   // 7. LIFECYCLE
   onMount(async () => {